Detach character stats from the static header
