+++ src/app/inventory/dimStats.scss
@@ -5,7 +5,6 @@
   opacity: 0.8;
   display: flex;
   flex-direction: row;
-  margin-top: 3px;
   justify-content: space-around;
 
   .phone-portrait & {
